pyproject.toml
[project]
name = "bank-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "openai>=2.8.1",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
